feat: add summary & description to various endpoints. Also update the atlas commit to use enhanced swagger instances.
Closes: #21
