+++ cabal.project
@@ -34,8 +34,8 @@ package strict-containers
 source-repository-package
   type: git
   location: https://github.com/geniusyield/atlas
-  tag: 2bb7431f273dd3b737fab4e7682e9457f3b10894
-  --sha256: sha256-z1jaOWqG3Dj0ZsLLlk4Pz7mOydhni+KOF96Z9suw9uc=
+  tag: 0cde387bc183d59509f4cd4479e112146087e043
+  --sha256: todo
 
 source-repository-package
   type: git
